feat: remove legacy autoloading feat: add zeitwerk loader by default refactor: extracted the caching mechanism into it own concern
