+++ lib/acts-as-taggable-on/taggable/collection.rb
