+++ lib/acts-as-taggable-on/taggable/core.rb
