+++ lib/acts-as-taggable-on/taggable/ownership.rb
