+++ lib/acts-as-taggable-on/taggable/related.rb
