+++ lib/acts-as-taggable-on/taggable/tag_list_type.rb
