+++ lib/acts-as-taggable-on/version.rb
