--- lib/acts_as_taggable_on.rb
@@ -1,6 +1,0 @@
-$stderr.puts <<EOF
-You are requiring 'acts_as_taggable_on' in the 'acts-as-taggable-on' gem.
-
-Please change your require to 'acts-as-taggable-on'.
-EOF
-require_relative 'acts-as-taggable-on'
