+++ lib/tasks/example/acts-as-taggable-on.rb.example
